requirements.txt
pandas>=1.0
numpy>=1.18
scipy>=1.4
matplotlib>=3.0
pyyaml>=5.1